changes in conflicted java file

+++ app/src/main/java/com/aniketjain/weatherapp/merge_conflict.java
@@ -1,7 +1,7 @@
 package com.aniketjain.weatherapp;
 
 public class merge_conflict {
-    int roll;
+    float roll;
     String name;
 
     public merge_conflict(int roll, String name) {
